requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113
torchvision==0.12.0+cu113
torchaudio==0.11.0
numpy==1.22.4
ray==2.3.0
ray[rllib]==2.3.0
docker==5.0.3
msgpack==1.0.5
msgpack_numpy==0.4.8
opencv-python==4.5.5.64
grpcio==1.54.0 
grpcio-tools==1.54.0 
protobuf==4.22.3
gymnasium==0.26.3
jsonlines==3.1.0
psutil==5.9.5
tqdm==4.66.1
huggingface_hub==0.17.1
filelock